working on the cart context

+++ react-auth/src/components/ProductItem.js
@@ -1,6 +1,7 @@
 import React, { useContext, useState } from 'react';
-import { addItemWL } from './addProductWL';
+import { useNavigate } from 'react-router-dom';
 import { WishlistContext } from '../context/WishlistContext';
+import { AuthContext } from '../context/AuthContext';
 import Cookies from 'universal-cookie'
 import axios from 'axios'
 const cookies = new Cookies()
@@ -9,40 +10,65 @@ export default function ProductItem({ brand, images, price, _id, name, title }) 
     const frontImages = images.slice(0, 2);
     const [hovered, setHovered] = useState(false);
     
-    const {isInWishlist, addToWishlist} = useContext(WishlistContext)
+    const {isAuthenticated} = useContext(AuthContext)
+    const {isInWishlist, addToWishlist,removeWishlist} = useContext(WishlistContext)
     const inWishlist = isInWishlist(_id)
 
+    const token = cookies.get("TOKEN")
+    const apiUrl = process.env.REACT_APP_API_URL
+    const navigate = useNavigate()
+    
     const handleItemWL = async() => {
 
-        const apiUrl = process.env.REACT_APP_API_URL
-        const token = cookies.get("TOKEN")
-        try{
-
-            if(inWishlist){
-                console.log(`Removing ${name} from wishlist database`)
-                await axios.delete(`${apiUrl}/api/wishlist/remove/${_id}`,{
-                    headers:{
-                        Authorization: `Bearer ${token}`
-                    }
-                })
-                console.log('removed from wishlist')
-            }
-            else{
-                console.log(`Adding ${name} to wishlist`)
-                await axios.post(`${apiUrl}/api/wishlist/add`,
-                    {productID: _id},
-                    {
+        if(isAuthenticated){
+            try{
+                if(inWishlist){
+                    console.log(`Removing ${name} from wishlist database`)
+                    await axios.delete(`${apiUrl}/api/wishlist/remove/${_id}`,{
                         headers:{
                             Authorization: `Bearer ${token}`
                         }
-                    }
-                )
+                    })
+                    removeWishlist(_id)
+                    console.log('removed from wishlist')
+                }
+                else{
+                    console.log(`Adding ${name} to wishlist`)
+                    await axios.post(`${apiUrl}/api/wishlist/add`,
+                        {productID: _id},
+                        {
+                            headers:{
+                                Authorization: `Bearer ${token}`
+                            }
+                        }
+                    )
+                    addToWishlist(_id)
+                    console.log('added to wishlist')
+                }
             }
+            catch(error){
+                console.error('Error with handleItemWL')
+                
+            }
+
         }
-        catch(error){
-            console.error('Error with handleItemWL')
+        else{
+            alert("Please log in to use the wishlist")
+            navigate("/contul-meu")
+            return
         }
 
+
+    }
+
+    const addItemCart = async() => {
+        try{
+
+        }
+        catch(error)
+        {
+
+        }
     }
 
 
@@ -68,6 +94,12 @@ export default function ProductItem({ brand, images, price, _id, name, title }) 
                     {title === "sales" && (
                         <p id='sale-price'>{Math.floor((price - price / 10) * 10) / 10}$</p>
                     )}
+
+                    <button 
+                        className='add-to-cart'
+                        onClick={addItemCart}>
+                        <img src='add-to-cart.png' alt="addToCart"/>
+                    </button>
                     <button 
                         className='favourite-item'
                         onClick={handleItemWL}>
